requirements.txt
torch
torchvision
torchaudio
transformers
unsloth
bitsandbytes
accelerate
xformers==0.0.29.post3
peft
trl==0.15.2
triton
cut_cross_entropy
unsloth_zoo
sentencepiece
protobuf
datasets>=3.4.1
huggingface_hub
hf_transfer
moviepy
opencv-python
openai
python-dotenv
fastapi
uvicorn
pyngrok
streamlit
requests
pillow
